frontend/requirements.txt
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
google-cloud-firestore==2.19.0
click==8.1.7
pandas==2.2.2
langchain==0.3.0
langchain-google-community==2.0.0
langchain-googledrive==0.3.0
langchain-core==0.3.0
langchain-google-vertexai==2.0.4
langchain-google-firestore==0.4.0
PyPDF2==3.0.1
python-docx==1.1.2
selenium==4.25.0
streamlit==1.39.0
unstructured==0.15.13
